"Customize" replaced by "See all the integrations" ✨
In this groundbreaking change, we not only renamed an option, but *also* replaced an image (upgraded to new high-definition and retina-ready). Now, when you’re in the <kbd>New Tab</kbd> view, you’ll be treated to a shiny new image (`gkc-newtab-integrations.png`), and you’ll be doing less “customizing” and more “integrating.”

Merging this will save an estimated 0.002 seconds of confusion per user. Thank me later. 🕰️🚀