Updates the GKC-Security document with detailed data security information. Corrects capitalization in the GitKraken Launchpad document.
Adds a new security documentation file outlining the
types of data collected, how it's secured in transfer
and storage, and where it is stored for various
GitKraken Desktop services.

Updates the security information document to use a table format for clarity and provides details on data collection, transfer security, storage location, and storage security for various GitKraken Desktop services.